Fix warning about frame_size_samples in MP3FrameInfo, by reducing it to uint16_t (which it shouldn't exceed anyway)

+++ MPEGAudioFile.h
@@ -11,7 +11,7 @@ struct MP3FrameInfo
 	size_t byte_position;       // Location of frame in the file
 	size_t frame_size_bytes;    // The size of the frame in bytes
 	size_t sample_position;	    // The first sample of this frame
-	size_t frame_size_samples;  // The size of the frame in samples
+	uint16_t frame_size_samples;  // The size of the frame in samples
 };
 
 // Functions:
